IA.I.R.E

+++ src/adc.h
@@ -67,11 +67,11 @@ extern "C" {
      */
     typedef struct
     {
-        uint16_t values[5];         /**< Save values sampled for ADC. */
-        uint16_t values_mv[5];      /**< Save values sampled for ADC in mV. */
+        uint32_t values[5];         /**< Save values sampled for ADC. */
+        uint32_t values_mv[5];      /**< Save values sampled for ADC in mV. */
         
-        uint16_t values_2[5][ADC_MAX_NUM_SAMPLE_PROM]; /**< Save values sampled for avergae. */
-        uint16_t values_2_count;    /**< Count samples for avergae. */
+        uint32_t values_2[5][ADC_MAX_NUM_SAMPLE_PROM]; /**< Save values sampled for avergae. */
+        uint32_t values_2_count;    /**< Count samples for avergae. */
         uint32_t values_2_prom[5];  /**< Save the average of sample. */
 
     } ADC_DATA;
